split appjs for agent test, mute problematic test case, clean code
