--- src/middleware/mtoattribute.js
@@ -1,11 +1,0 @@
-'use strict';
-
-module.exports= class MTOattribute{
-    constructor(createtime,changetime,modifytime,size,name){
-        this.changetime = changetime;
-        this.modifytime = modifytime;
-        this.createtime = createtime;
-        this.size = size;
-        this.name = name;
-    }
-}
