--- src/middleware/mtopermission.js
@@ -1,8 +1,0 @@
-'use strict';
-module.exports = class MTOpermission{
-    constructor(readlist,writelist,owner){
-        this.readlist = readlist;
-        this.writelist = writelist;
-        this.owner = owner;
-    }
-}
