--- src/middleware/treebuilder.js
@@ -1,94 +1,0 @@
-const uuid = require('node-uuid');
-const globby = require('globby');
-var fs = require('fs');
-var xattr = require('fs-xattr');
-var dmap1 = new Map();
-var helper = require('../middleware/tools');
-var adapter = require('../middleware/adapter');
-var exif = require('../middleware/exif');
-var MTOpermission = require('../middleware/mtopermission');
-var MTOattribute = require('../middleware/mtoattribute');
-var MTObj = require('../middleware/memtree');
-var debug=false
-
-function commoncheck(f){
-  debug && console.log(f)
-  helper.tattoo(f);
-  //console.log(9991)
-  var fstat=fs.statSync(f);
-  var uid = xattr.getSync(f,'user.uuid').toString('utf-8');
-  var readlist = xattr.getSync(f,'user.readlist').toString('utf-8').split(',');
-  var writelist = xattr.getSync(f,'user.writelist').toString('utf-8').split(',');
-  var owner = xattr.getSync(f,'user.owner').toString('utf-8').split(',');
-  var type = xattr.getSync(f,'user.type').toString('utf-8');
-  try{
-    var hash=xattr.getSync(f,'user.hash').toString('utf-8');
-  }
-  catch(e){
-    var hash=""
-  }
-  var createtime = fstat.birthtime;
-  var changetime = fstat.ctime;
-  var modifytime = fstat.mtime;
-  var size = fstat.size;
-  var mtobj = adapter.treebuilder(uid,readlist,writelist,owner,type,createtime,changetime,modifytime,size,f,'',hash,'');
-  
-  if(dmap.has(f.substr(0,f.lastIndexOf('/')))){
-    var parent = dmap.get(f.substr(0,f.lastIndexOf('/')));
-    mtobj.parent=parent;
-  }
-  if (fstat&&fstat.isDirectory()){ 
-    if(!memt.has(mtobj.uid)){
-      var mtop=new MTOpermission(mtobj.readlist,mtobj.writelist,mtobj.owner);
-      var mtoa= new MTOattribute(mtobj.createtime,mtobj.changetime,mtobj.modifytime,mtobj.size,mtobj.path.substr(mtobj.path.lastIndexOf('/')+1));
-      var memobj = new MTObj(mtobj.uid,mtobj.type,mtobj.parent,[],mtobj.path,mtop,mtoa,mtobj.hash);
-      memt.add(mtobj.uid,memobj);
-      dmap.set(mtobj.path,mtobj.uid);
-    }
-  }
-  else{
-    if(!memt.has(mtobj.uid)){
-      var mtop=new MTOpermission(mtobj.readlist,mtobj.writelist,mtobj.owner);
-      var mtoa= new MTOattribute(mtobj.createtime,mtobj.changetime,mtobj.modifytime,mtobj.size,mtobj.path.substr(mtobj.path.lastIndexOf('/')+1));
-      var memobj = new MTObj(mtobj.uid,mtobj.type,mtobj.parent,[],mtobj.path,mtop,mtoa,mtobj.hash,'');
-      memt.add(mtobj.uid,memobj);
-      //var a =helper.pastedetail(mtobj.path,mtobj.uid);
-      exif.save(memt.get(mtobj.uid));
-      //console.log(msg.uid);
-      //console.log(msg.path);
-    }
-  }
-
-  if(dmap.has(f.substr(0,f.lastIndexOf('/')))){
-    memt.addchild(mtobj.parent, memt.get(mtobj.uid))
-  }
-  else{
-    memt.setroot(mtobj.uid)
-  }
-}
-
-function cronjob(tpath){
-  var newlist = globby.sync([tpath]);
-  newlist.forEach(function(f){
-    // console.log("----")
-    console.log(f)
-    // console.log("----")
-    var tf=f.split('/');
-    if(f==='/data/fruitmix'){
-      try{
-        var puuid=xattr.getSync(f,'user.uuid');
-      }
-      catch(e){
-        var puuid=uuid.v4()
-        xattr.setSync(f,'user.uuid',puuid);
-      }
-      memt.setroot(puuid);
-    }
-    if(f==='/data/fruitmix'||helper.contains(tf,'library')||helper.contains(tf,'drive')){
-      commoncheck(f);
-    }
-  });
-}
-
-exports.checkall = cronjob;
-exports.docheck = commoncheck;
